Split up cli_test.ts
